all tests apart from flush() pass
